refactor: move generation functions to generate.py and update imports
